requirements.txt
# Core dependencies
google-generativeai>=0.8.0
mcp>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Terminal UI
rich>=13.0.0
prompt-toolkit>=3.0.0

# Async support
aiofiles>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
